:sparkles: feat(explorer): add Act UI for batch file operations (mc/mx/md)
- Add dot.board.Act component for input-preview UI pattern
- Implement mc (copy selected), mx (move selected), md (delete selected) with preview
- Auto-trigger mc/mx when pressing p with multiple selections in copy/cut mode
- Add get_common_ancestor_path for computing shared parent directory
- Add get_selected_nodes_toplevel for efficient batch selection handling
- Fix unused variables and nil checks in tree.lua
- Remove unnecessary usize casts in Rust search module
- Replace vim.endswith with string:sub(-1) to avoid spell checker warnings

+++ rust/yoz/src/search/search_in_lines_regex.rs
@@ -32,11 +32,11 @@ pub fn search_in_lines_regex(
             let start_pos = mat.start();
             let end_pos = mat.end();
             let line_num = locate_line(line_offsets, start_pos);
-            if line_num == 0 || (line_num as usize) > buffer.line_count() {
+            if line_num == 0 || line_num > buffer.line_count() {
                 continue;
             }
 
-            let line_index = (line_num - 1) as usize;
+            let line_index = line_num - 1;
             let line_start_pos = line_offsets[line_index];
             let relative_start = start_pos.saturating_sub(line_start_pos);
             let relative_end = end_pos.saturating_sub(line_start_pos);
